refactor(#140 WI-5): extract Azw3ReaderChrome + jump helpers into a sibling file
Pure, behaviour-preserving move ahead of WI-6's wiring so that WI's diff and
audit stay about behaviour, not relocation noise. Azw3ReaderActivity.kt was
569 lines, past the ~300-line guideline.

Moved verbatim into the SAME package (com.vreader.app.reader), so no import
anywhere changed — main or androidTest:
  - Azw3ReaderChrome (internal @Composable)
  - Azw3NotesBottomChrome (private @Composable, moves with its only caller)
  - azw3JumpResult / azw3JumpDecision (+ their section comment)

Azw3ReaderActivity.kt's diff is deletions only (0 additions, 147 deletions):
the four symbols plus the nine imports they alone used. 569 -> 422 lines;
the new file is 186.

No test file changed. Azw3BookmarkNavTest / Azw3ReaderChromeUiTest /
SearchHiddenOnPdfAzw3Test / BookmarkHostWiringTest all reference the moved
symbols unqualified and compile untouched.

Refs #140
